refactor: delete util
