+++ src/main/java/Calculator/Enums/OperatorEnum.java
@@ -2,7 +2,7 @@ package Calculator.Enums;
 
 import java.util.function.BiFunction;
 
-public enum Operator {
+public enum OperatorEnum {
     PLUS("+", (a, b) -> a + b),
     MINUS("-", (a, b) -> a - b),
     MULTIPLE("*", (a, b) -> a * b),
@@ -11,7 +11,7 @@ public enum Operator {
     private String symbol;
     private BiFunction<Integer, Integer, Integer> function;
 
-    Operator(String symbol, BiFunction<Integer, Integer, Integer> function) {
+    OperatorEnum(String symbol, BiFunction<Integer, Integer, Integer> function) {
         this.symbol = symbol;
         this.function = function;
     }
